# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeAccountStatus } from './employeeAccountStatus';

export interface Employee {
  id: number;
  name: string;
  role: string;
  initials: string;
  department: string;
  /** @nullable */
  email: string | null;
  isAdmin: boolean;
  accountStatus: EmployeeAccountStatus;
}