# lib/api-zod/src/generated/types/employeeInvitationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInvitationInput {
  email: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  department: string;
  employeeId?: number;
}